backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
transformers==4.46.3
torch==2.3.1
bitsandbytes==0.43.3
accelerate==0.33.0
streamlit==1.36.0
requests==2.32.3
pandas==2.2.2
cachetools==5.3.3
scipy==1.13.1
plotly==5.18.0

# 🔥 ADD THESE (for RAG)
faiss-cpu==1.8.0
sentence-transformers==3.0.1